+++ packages/authhero/CHANGELOG.md
@@ -1,5 +1,17 @@
 # authhero
 
+## 0.81.0
+
+### Minor Changes
+
+- Add a cloudflare adapter
+
+## 0.80.0
+
+### Minor Changes
+
+- Use correct email template
+
 ## 0.79.0
 
 ### Minor Changes